+++ src/integrations/ninja-forms.php
@@ -102,6 +102,8 @@ function standardize_location_data( string $value, string $key ): string {
 		$suffixes = [
 			'Dr'    => 'Drive',
 			'Ln'    => 'Lane',
+			'Lm'	=> 'Lane',
+			'Lsne'	=> 'Lane',
 			'Tr'    => 'Trail',
 			'Cir'   => 'Circle',
 			'Av'    => 'Avenue',
@@ -111,6 +113,7 @@ function standardize_location_data( string $value, string $key ): string {
 			'Pkwy'  => 'Parkway',
 			'Bl'    => 'Boulevard',
 			'St'    => 'Street',
+			'Ter'   => 'Terrace',
 			'Terr'  => 'Terrace',
 			'Pl'    => 'Place',
 			'Wy'    => 'Way',
@@ -128,6 +131,9 @@ function standardize_location_data( string $value, string $key ): string {
 			'Nw'	=> 'Northwest',
 			'E'		=> 'East',
 			'W'		=> 'West',
+			'Cv'	=> 'Curve',
+			'Crv'	=> 'Curve',
+			'Curv'	=> 'Curve',
 
 		];